Renamed package name except for VM testbase conversion comment
