+++ test/hotspot/jtreg/vmTestbase/vm/mlvm/hiddenloader/func/findByName/Test.java
@@ -37,18 +37,18 @@
  *          /test/lib
  *
  * @comment build test class and indify classes
- * @build vm.mlvm.anonloader.func.findByName.Test
+ * @build vm.mlvm.hiddenloader.func.findByName.Test
  * @run driver vm.mlvm.share.IndifiedClassesBuilder
  *
- * @run main/othervm vm.mlvm.anonloader.func.findByName.Test
+ * @run main/othervm vm.mlvm.hiddenloader.func.findByName.Test
  */
 
-package vm.mlvm.anonloader.func.findByName;
+package vm.mlvm.hiddenloader.func.findByName;
 
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodHandles.Lookup;
 
-import vm.mlvm.anonloader.share.AnonkTestee01;
+import vm.mlvm.hiddenloader.share.AnonkTestee01;
 import vm.mlvm.share.MlvmTest;
 import vm.share.FileUtils;
