+++ test/hotspot/jtreg/vmTestbase/vm/mlvm/hiddenloader/share/HiddenkTestee02.java
@@ -21,7 +21,7 @@
  * questions.
  */
 
-package vm.mlvm.anonloader.share;
+package vm.mlvm.hiddenloader.share;
 
 public class HiddenkTestee02 extends HiddenkTestee01 {
     public HiddenkTestee02() {}
